Use doTransfer for revokeTokenGrant
This lets us reuse the existing transfer functionality which MiniMe
provides. I use `doTransfer` and not `transferFrom` to go around the
vesting functionality. The only downside here is that we could only
revoke grants after the sale is over once we have a different controller
in place. This is fine though since the cliffs of all of the tokens we
are granting should be at least two months after a controller that
supports transfers is released.

Fixes #66